restored old HistHelper.py again
